Enable image scaling.
This patch disables the DocBook "ignore.image.scaling" feature. It
allows to specify image scaling options, e.g. to scale an image to
a certain percentage of the available space. Used to provide high
resolution images, but automatically scale them down to the
available space (both in HTML and PDF outputs).

References:
http://docbkx-tools.sourceforge.net/docbkx-maven-plugin/generate-html-mojo.html#ignoreImageScaling
http://docbkx-tools.sourceforge.net/docbkx-maven-plugin/generate-pdf-mojo.html#ignoreImageScaling
http://docbook.sourceforge.net/release/xsl/1.78.1/doc/html/ignore.image.scaling.html
Change-Id: I2bd8e055c76e016de98cab457b20db2b209d4d1e
Signed-off-by: Jan Hilberath <jan@midokura.com>