Kilo LBaaS configuration
Starting in Kilo, LBaaS is implemented by the MidoNet plugin
as an advanced service driver instead of a top level plugin.

By implementing as an advanced service driver, it follows the
recommended model of advanced service projects, and provides
better integration with other OpenStack components such as Horizon.

The section added in this patch describes the additional
configuration required to enable the driver.  Without this
configuration, the Neutron server would not start successfully.

Change-Id: If95eac94d9f995fbaf6494ad8c444a39f89eaa30
Implements: MND-36
Signed-off-by: Jan Hilberath <jan@midokura.com>