QSG: Fix Cassandra init script on RHEL 7
During installation of the Cassandra RPM, the `/var/run/cassandra`
directory is created, but because it is located on a temporary file
system it will be lost after system reboot. As a result it is not
possible to stop or restart the Cassandra service anymore.

This patch set adds instructions to edit the `/etc/init.d/cassandra`
file to create the directory on service start.

Fixes: MND-41

Change-Id: I6884ade6590d1c294d6d058637ddd09e6bb5a8d9
Signed-off-by: Jan Hilberath <jan@midokura.com>
(cherry picked from commit 4ddab90749318eec491e8e95768ff68a98fb85cc)