QSGs: Change initial network (and MPR) creation and BGP setup order
The Quick Start Guides assume the MidoNet Provider Router (MPR) to be
already created when setting up BGP connectivity.

But the MRP is automatically created when creating an external network
in Neutron, which came after the BGP setup.

This patch set changes the order, adds an example Neutron command for
the external network creation, and adds a short note about the automatic
MPR creation.

Change-Id: I25717c9014011284301a7925149922c3b0ee4cb8
Fixes: MND-113
Signed-off-by: Jan Hilberath <jan@midokura.com>