QSGs: Change initial network creation and Edge Router / BGP setup order
The Quick Start Guides assumes an external network to be already created
when setting up and Edge Router and configuring BGP connectivity.

But the external network creation section came after the Edge Router and
BGP setup sections.

This patch set changes the order and adds an example Neutron command for
the external network creation.

Change-Id: I354386d3cf8da6e8bd75fff7ae946ae656d32988
Fixes: MND-113
Signed-off-by: Jan Hilberath <jan@midokura.com>