OP Guide: Update BGP configuration
This patch set updates the Operation Guide's BGP uplink
configuration section according to the one currently
being used in the Quick Start Guides.

The "Adding a second session on the same router port"
and "Adding BGP sessions to a second router port"
sections (which do not exist in the Quick Start Guides)
have been moved to separate source files as is.

Change-Id: Id4a727c5d0057e4eb9efccff8bd12be8f0646b8a
Ref: MND-130
Signed-off-by: Jan Hilberath <jan@midokura.com>