Removed cleart/sett from BGP configuration
With MidoNet v5.0 the Edge Router is owned by the admin tenant, it is
not necessary anymore to clear the tenant in midonet-cli as it was in
previous versions with the MidoNet Provider Router.

Additionally, in the Japanese version, this patch set replaces the
term "MidoNet Provider Router" with "Edge Router".

Change-Id: I63d02b71ad7023ab8c3efcfd795e81d25f5fbe76
Fixes: MND-147
Signed-off-by: Jan Hilberath <jan@midokura.com>