Share single BGP configuration source between OP Guide and QSGs
In the past we were maintaining two versions of the BGP configuration
sources, one in the Operations Guide, the other in the Quick Start
Guide source tree.

To make maintenance easier this patch set creates symlinks from the
QSG source tree to the OP Guide sources, so that only a single source
file exists.

Change-Id: Iad88d5ef1a2080ec631f6f3d4532df66865a7e32
Ref: MND-148
Signed-off-by: Jan Hilberath <jan@midokura.com>