Service containers at the agent
This includes documentation for troubleshooting the service
containers at the agent including the following.

A checklist for creating, updating and deleting containers,
including the typical log messages an operator should expect in
the agent log.

Information on the containers operations log, and how the
operator can use it to troubleshoot running containers.

Information on configuring the container logging, related to how
the logging from the processes running inside the container is
appended to the agent log, how the operator and identify it, and
how it can enable/disable it.

Ref: MI-62

Change-Id: I1a8018cec0913c1d18b6534849faf57f24e5cf85
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Jan Hilberath <jan@midokura.com>