QSGs: Don't disable Keystone admin_token_auth
The Liberty QSGs were missing to mention that Keystone admin_token_auth
must not be disabled as mentioned in the upstream docs. (MidoNet API
will fail to authenticate if disabled.)

This patch set adds the missing note to the Liberty QSG, and an
additional step to verify the configuration to all QSGs.

Change-Id: Ie08838b94f8f8843a99ef454a415b506f3fdecc6
Fixes: MND-59
Signed-off-by: Jan Hilberath <jan@midokura.com>