Static documentation Kuryr
This commit introduces a full section for Kuryr (tech-preview).
As opposite to the rest of the documentation here, this is intended to
be on OpenStack upstream in the future (maybe we'll keep a fork) and
this is why the style and the way to generate it is different.

The way to generate this documentation is via `tox -edocs` in the
midonet/kuryr project. The generated results are being copy-pasted on
this project.

Maybe in a future we can find a way to synchronize efforts without
manual steps....

Change-Id: I911904ce4d2a7358e39f58cac9a9e360bffaa53f